% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{add_weather_data}
\alias{add_weather_data}
\title{Fügt data frame mit COVID Daten die durchschnittliche Temperatur (für alle verfügbaren Wetterstationen in entsprechendem Landkreis) des Entsprechenden Tages hinzu.}
\usage{
add_weather_data(cov_data)
}
\arguments{
\item{cov_data}{Data frame mit COVID-Daten. Muss Spalte "IdLandkreis", sowie "Meldedatum" enthalten.}
}
\value{
data frame mit neuer Spalte "Temperatur"
}
\description{
Fügt data frame mit COVID Daten die durchschnittliche Temperatur (für alle verfügbaren Wetterstationen in entsprechendem Landkreis) des Entsprechenden Tages hinzu.
}
