% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{append_report_duration}
\alias{append_report_duration}
\title{Fügt Spalte "date_diff" an data frame an, welche die Anzahl Tage zwischen Infektion und Meldung beim Gesundheitsamt beschreibt.}
\usage{
append_report_duration(data)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}
}
\value{
Data frame mit gesuchte Daten und beschreibendem Flag
}
\description{
Fügt Spalte "date_diff" an data frame an, welche die Anzahl Tage zwischen Infektion und Meldung beim Gesundheitsamt beschreibt.
}
\examples{
append_report_duration(df)
}
