% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidproject-package.R
\docType{package}
\name{covidproject-package}
\alias{covidproject}
\alias{covidproject-package}
\title{covidproject: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
