% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Downloader.R
\name{data_downloader}
\alias{data_downloader}
\title{Läd die csv Datei mit COVID-Daten herunter.}
\usage{
data_downloader(path)
}
\arguments{
\item{path}{Pfad in dem die csv Datei gespeichtert werden soll.}
}
\description{
Läd die csv Datei mit COVID-Daten herunter.
}
