% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{download_population_data}
\alias{download_population_data}
\title{lädt Daten zu Bevölkerungszahlen pro Landkreis herunter, welche zur Berechnung der Inzidenzen benötigt werden}
\usage{
download_population_data()
}
\description{
lädt Daten zu Bevölkerungszahlen pro Landkreis herunter, welche zur Berechnung der Inzidenzen benötigt werden
}
