% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{download_weather_data}
\alias{download_weather_data}
\title{Lädt historische Klimadaten von Deutschem Wetterdienst herunter}
\usage{
download_weather_data()
}
\description{
Lädt historische Klimadaten von Deutschem Wetterdienst herunter
}
