% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_by_age_group}
\alias{filter_by_age_group}
\title{Filtert nach Altersgruppen im data frame, die zwischen der spezifizierten Start- und Endaltersgruppe liegen.}
\usage{
filter_by_age_group(data, age_group_start = NA, age_group_end = NA)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{age_group_start}{String mit Startaltersgruppe (z. B. "A00")}

\item{age_group_end}{String mit Startaltersgruppe (z. B. "A59")}
}
\value{
gefiltertes data frame
}
\description{
Filtert nach Altersgruppen im data frame, die zwischen der spezifizierten Start- und Endaltersgruppe liegen.
}
