% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_by_date}
\alias{filter_by_date}
\title{Filtert nach Einträgen im data frame, die zwischen dem spezifizierten Start- und Enddatum liegen.}
\usage{
filter_by_date(data, date_start, date_end)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{date_start}{Startdatum als String ("YYYY/MM/DD")}

\item{date_end}{Enddatum als String ("YYYY/MM/DD")}
}
\value{
gefiltertes data frame
}
\description{
Filtert nach Einträgen im data frame, die zwischen dem spezifizierten Start- und Enddatum liegen.
}
