% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_correlation_for_incidence_pairs}
\alias{get_correlation_for_incidence_pairs}
\title{Erzeugt data frame, welches die Correlation zwischen Landkreispaaren bezüglich derer Inzidenz beschreibt}
\usage{
get_correlation_for_incidence_pairs(incidence_data)
}
\arguments{
\item{incidence_data}{Data frame mit Inzidenz Daten. Erzeugbar durch get_incidence_per_district(...)}
}
\value{
Data frame mit gesuchte Daten
}
\description{
Erzeugt data frame, welches die Correlation zwischen Landkreispaaren bezüglich derer Inzidenz beschreibt
}
\examples{
get_correlation_for_incidence_pairs(incidence.df)
}
