% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_fallsterblichkeit_overall}
\alias{get_fallsterblichkeit_overall}
\title{Erzeugt Daten zu Fallsterblichkeit in ganz Deutschland abhängig von Altersgruppen}
\usage{
get_fallsterblichkeit_overall(data, age_group_start = NA, age_group_end = NA)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{age_group_start}{String mit Startaltersgruppe (z. B. "A00")}

\item{age_group_end}{String mit Startaltersgruppe (z. B. "A59")}
}
\value{
Data frame mit gesuchte Daten und beschreibendem Flag
}
\description{
Erzeugt Daten zu Fallsterblichkeit in ganz Deutschland abhängig von Altersgruppen
}
\examples{
get_fallsterblichkeit_overall(df, age_group_start = "A00", age_group_end = "A59")
}
