% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_federal_state_by_district_id}
\alias{get_federal_state_by_district_id}
\title{Gibt zurück, in welchem Bundesland sich ein Landkreis befindet.}
\usage{
get_federal_state_by_district_id(district_id)
}
\arguments{
\item{district_id}{IdLandkreis (Amtlicher Gemeindeschlüssel, AGS) des Landkreises}
}
\value{
Bundeslandname
}
\description{
Gibt zurück, in welchem Bundesland sich ein Landkreis befindet.
}
