% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_incidence_per_district}
\alias{get_incidence_per_district}
\title{Erzeugt Daten zu Inzidenzen pro District.}
\usage{
get_incidence_per_district(
  data,
  age_group_start = "A00",
  age_group_end = "A80",
  district = NA,
  date_start = NA,
  date_end = NA,
  incidence_days = 7
)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{age_group_start}{String mit Startaltersgruppe (z. B. "A00")}

\item{age_group_end}{String mit Startaltersgruppe (z. B. "A59")}

\item{district}{Einzelner String oder Vektor aus Strings mit angefragten Landkreisnamen (beginnend mit z. B. "LK" f端r Landkreis und "SK" f端r Stadtkreis). Falls district NA ist werden Daten f端r alle Landkreise erzeugt.}

\item{date_start}{Startdatum als String ("YYYY/MM/DD")}

\item{date_end}{Enddatum als String ("YYYY/MM/DD")}

\item{incidence_days}{Zahl der Tage zur Berechnung der Inzidenz, z. B. incidence_days = 7 f端r 7-Tage-Inzidenz}
}
\value{
Data frame mit gesuchte Daten
}
\description{
Erzeugt Daten zu Inzidenzen pro District.
}
\examples{
get_incidence_per_district(df, age_group_start = "A00", age_group_end = "A59", district = "LK Karlsruhe", date_start = "2021/05/01", date_end = "2021/06/01")
}
