% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_infections_overall}
\alias{get_infections_overall}
\title{Erzeugt Daten zu Infektionen mit bekanntem Krankheitsbeginndatum in ganz Deutschland}
\usage{
get_infections_overall(
  data,
  age_group_start = NA,
  age_group_end = NA,
  date_start = NA,
  date_end = NA
)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{age_group_start}{String mit Startaltersgruppe (z. B. "A00")}

\item{age_group_end}{String mit Startaltersgruppe (z. B. "A59")}

\item{date_start}{Startdatum als String ("YYYY/MM/DD")}

\item{date_end}{Enddatum als String ("YYYY/MM/DD")}
}
\value{
Data frame mit gesuchte Daten und beschreibendem Flag
}
\description{
Erzeugt Daten zu Infektionen mit bekanntem Krankheitsbeginndatum in ganz Deutschland
}
\examples{
get_infections_overall(df, age_group_start = "A00", age_group_end = "A59", date_start = "2021/05/01", date_end = "2021/06/01")
}
