% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_recovered_per_district}
\alias{get_recovered_per_district}
\title{Erzeugt Daten zu Genesenenf채llen pro Landkreis}
\usage{
get_recovered_per_district(
  data,
  age_group_start = NA,
  age_group_end = NA,
  district = NA,
  date_start = NA,
  date_end = NA
)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}

\item{age_group_start}{String mit Startaltersgruppe (z. B. "A00")}

\item{age_group_end}{String mit Startaltersgruppe (z. B. "A59")}

\item{district}{Einzelner String oder Vektor aus Strings mit angefragten Landkreisnamen (beginnend mit z. B. "LK" f체r Landkreis und "SK" f체r Stadtkreis)}

\item{date_start}{Startdatum als String ("YYYY/MM/DD")}

\item{date_end}{Enddatum als String ("YYYY/MM/DD")}
}
\value{
Data frame mit gesuchte Daten und beschreibendem Flag
}
\description{
Erzeugt Daten zu Genesenenf채llen pro Landkreis
}
\examples{
get_recovered_per_district(df, age_group_start = "A00", age_group_end = "A59", district = c("LK Karlsruhe", "SK Karlsruhe"), date_start = "2021/05/01", date_end = "2021/06/01")
get_recovered_per_district(df, district = c("LK Karlsruhe", "SK Karlsruhe"), date_start = "2021/05/01", date_end = "2021/06/01")
get_recovered_per_district(df, district = c("LK Karlsruhe", "SK Karlsruhe"))
}
