% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{get_unclean_data}
\alias{get_unclean_data}
\title{Erzeugt data frame bestehend aus "unsauberen" Daten, also Unklarheiten und fehlerhaften Einträgen.}
\usage{
get_unclean_data(data)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}
}
\value{
Data frame mit gesuchte Daten
}
\description{
Erzeugt data frame bestehend aus "unsauberen" Daten, also Unklarheiten und fehlerhaften Einträgen.
}
\examples{
get_unclean_data(df)
}
