% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_district_map}
\alias{plot_district_map}
\title{Erzeugt, bzw. Speichert Plot von Karte Deutschlands, welche Landkreise darstellt}
\usage{
plot_district_map(data, save_in = NA, file_name = NA)
}
\arguments{
\item{data}{Data frame mit Daten zu Todesfällen, Infektionen, oder Genesenenzahlen}

\item{save_in}{Pfad zu Verzeichnis, in welchem der Plot gespeichert werden soll. Wenn save_in NA ist, wird der Plot direkt angezeigt}

\item{file_name}{Name der zu speichernden Bilddatei}
}
\value{
NULL, bzw. Plot
}
\description{
Erzeugt, bzw. Speichert Plot von Karte Deutschlands, welche Landkreise darstellt
}
\examples{
plot_district_map(df, save_in = "some/path", file_name = "map_plot")
}
