% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_function}
\alias{plot_function}
\title{Erzeugt, bzw. Speichert Plot aus data frame. Kann als Barchart dargestellt werden, oder als Linien (für Zeitreihen)}
\usage{
plot_function(data, add_weather = FALSE, scaling_coeff = 700, save_in = NA)
}
\arguments{
\item{data}{Data frame mit Daten aus get_... Methoden. Sollte ein beschreibendes "flag" als Attribut enthalten.}

\item{add_weather}{Boolean. Gibt an, ob zusätzlich bei einer Zeitreihe die Durschnittstemperatur des entsprechenden Datums geplottet werden soll.}

\item{scaling_coeff}{Zahl, um die Skalierung der beiden Achsen bei Wetter-Plots anzupassen.}

\item{save_in}{Pfad zu Verzeichnis, in welchem der Plot gespeichert werden soll. Wenn save_in NA ist, wird der Plot direkt angezeigt}
}
\value{
NULL, bzw. Plot
}
\description{
Erzeugt, bzw. Speichert Plot aus data frame. Kann als Barchart dargestellt werden, oder als Linien (für Zeitreihen)
}
\examples{
plot_function(df, add_weather=TRUE, save_in="some/path")
}
