% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_incidence_correlations_barchart}
\alias{plot_incidence_correlations_barchart}
\title{Erzeugt, bzw. Speichert Barchart von höchsten Inzidenz-Correlations-Paaren aus data frame.}
\usage{
plot_incidence_correlations_barchart(
  correlations_data,
  top = 10,
  save_in = NA,
  file_name = NA
)
}
\arguments{
\item{correlations_data}{Data frame mit Daten zu Inzidenz-Correlations-Paaren von Landkreisen.}

\item{top}{Anzahl der Paare, die betrachtet werden sollen. "top = 10" bedeutet die 10 Landkreispaare mit der höchsten Correlation zwischen Inzidenzen}

\item{save_in}{Pfad zu Verzeichnis, in welchem der Plot gespeichert werden soll. Wenn save_in NA ist, wird der Plot direkt angezeigt}

\item{file_name}{Name der zu speichernden Bilddatei}
}
\value{
NULL, bzw. Plot
}
\description{
Erzeugt, bzw. Speichert Barchart von höchsten Inzidenz-Correlations-Paaren aus data frame.
}
\examples{
plot_incidence_correlationsbarchart(df, top = 10, save_in = "some/path", file_name = "correlation_matrix_plot")
}
