% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_incidence_correlations_matrix}
\alias{plot_incidence_correlations_matrix}
\title{Erzeugt, bzw. Speichert Correlationsmatrix-Plot aus data frame.}
\usage{
plot_incidence_correlations_matrix(
  correlations_data,
  districts = NA,
  save_in = NA,
  file_name = NA
)
}
\arguments{
\item{correlations_data}{Data frame mit Daten zu Inzidenz-Correlations-Paaren von Landkreisen.}

\item{districts}{Vektor aus Strings mit Landkreisnamen, die in der Matrix dargestellt werden sollen (zur Übersichtlichkeit).}

\item{save_in}{Pfad zu Verzeichnis, in welchem der Plot gespeichert werden soll. Wenn save_in NA ist, wird der Plot direkt angezeigt}

\item{file_name}{Name der zu speichernden Bilddatei}
}
\value{
NULL, bzw. Plot
}
\description{
Erzeugt, bzw. Speichert Correlationsmatrix-Plot aus data frame.
}
\examples{
plot_incidence_correlations_matrix(df, districts = c("LK Kiel", "LK Karlsruhe", "SK Karlsruhe"), save_in = "some/path", file_name = "correlation_matrix_plot")
}
