% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plot_map}
\alias{plot_map}
\title{Erzeugt aus data frame einen Plot von Deutschlandkarte mit Landkreisen und Bundesländern}
\usage{
plot_map(cov_data, plot_title, legend_title)
}
\arguments{
\item{cov_data}{Ein data frame mit den Spalten "IdLandkreis" und "datapoints". "datapoints" wird dann farblich pro Landkreis dargestellt}

\item{plot_title}{Titel des Plots}

\item{legend_title}{Titel der Legende}
}
\value{
Plot objekt mit Deutschlandkarten-Plot
}
\description{
Erzeugt aus data frame einen Plot von Deutschlandkarte mit Landkreisen und Bundesländern
}
\examples{
plot_map(incidence.df, plot_title = "Beispiel", legend_title = "noch ein Beispiel")
}
