% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{predict_future_progress}
\alias{predict_future_progress}
\title{Nutzt Prophet um Vorhersagen zu zukünftigen Entwicklungen (z. B. Infektionen) zu machen}
\usage{
predict_future_progress(data, future_days = 730)
}
\arguments{
\item{data}{Data frame mit Daten zu Todesfällen, Infektionsfällen, oder Genesenenfällen}

\item{future_days}{Anzahl Tage, die in die Zukunft vorhergesagt werden soll}
}
\value{
NULL, bzw. Plot mit Vorhersage
}
\description{
Nutzt Prophet um Vorhersagen zu zukünftigen Entwicklungen (z. B. Infektionen) zu machen
}
\examples{
predict_future_progress(df)
}
