% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{remove_unclean_data}
\alias{remove_unclean_data}
\title{Reinigt data frame von "unsauberen" Daten, also Unklarheiten und fehlerhaften Einträgen.}
\usage{
remove_unclean_data(data)
}
\arguments{
\item{data}{Data frame mit COVID-Daten}
}
\value{
Gesäubertes data frame
}
\description{
Reinigt data frame von "unsauberen" Daten, also Unklarheiten und fehlerhaften Einträgen.
}
\examples{
remove_unclean_data(df)
}
