% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidproject-package.R
\docType{package}
\name{covidproject-package}
\alias{covidproject}
\alias{covidproject-package}
\title{covidproject: Analysis on COVID Data for Germany}
\description{
Analysis on COVID Data for Germany
}
\details{
doku und so
}
\section{tolle section}{
NA
}

\keyword{internal}
