% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidproject-package.R
\docType{package}
\name{covidproject-package}
\alias{covidproject}
\alias{covidproject-package}
\title{covidproject: Analysis of COVID Data for Germany}
\description{
This package gives tools to download, analyse and visualize data of the COVID19 virus for Germany.
}
\details{
Das COVID-Project ist ein Package, welches dem Nutzer aktuelle Daten des Robert-Koch-Instituts zur Coronapandemie zur Verfügung stellt. Die Datentabelle wird anhand der Anforderungen des Nutzers neu gefiltert und Daten werden aggregiert, um dem Nutzer spezifische Fragen zu den Coronazahlen zu beantworten. Zudem können die Daten in verschiedensten Plots dargestellt werden.
}
\keyword{internal}
